lwm2m: fix bug observe after reboot - tests refactoring 3
